Fix directory tracking interfering with projectile (Emacs)
Tracking directories is there in order for shells inside Emacs to pick up on
what the current directory is. This does some weird printing that I don't
understand, and interferes when Emacs runs shell commands and parses the
output. This completely brakes projectile mode in Emacs.

The fix was apparently to not attempt to track the directory if the terminal is
too dumb.